requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121torchtorchvisiontorchaudioDiffusers
diffusers
transformers
accelerate
pillow